% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{add_data_column}
\alias{add_data_column}
\title{Internal Helper to Add Data Column from Excel Files}
\usage{
add_data_column(df, column_name, sheets, sheets_regex)
}
\arguments{
\item{df}{A tibble.}

\item{column_name}{The name of the new column to be added.}

\item{sheets}{A character vector specifying which sheets to extract from the Excel files.}

\item{sheets_regex}{A regular expression pattern to match against sheet names in the
Excel files.}
}
\value{
A tibble with the original data and an additional column containing data from
the specified Excel sheets.
}
\description{
The \code{add_data_column} function is an internal helper function that adds a column
to a tibble containing data extracted from Excel files.
}
\keyword{internal}
