% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{add_form_name_version}
\alias{add_form_name_version}
\title{Add a form name/version to a file_tibble}
\usage{
add_form_name_version(df, new_form_name_version)
}
\arguments{
\item{df}{A tibble containing the data to be reformed.}

\item{new_form_name_version}{A string containing the name_version of a form template. Format is "name.v1.0.0".}
}
\value{
A tibble with a nested metadata column that contains the new_form_name_version
}
\description{
This is a helper function that is only necessary for forms that don't already have the form name/verison in the top left corner of the first sheet.
}
