% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_form_to_file_name.R
\name{add_form_to_string}
\alias{add_form_to_string}
\title{Add form version to file name}
\usage{
add_form_to_string(file_name, form_name_version)
}
\arguments{
\item{file_name}{The name of the file to which the version will be appended.}

\item{form_name_version}{The version string to append to the file name.}
}
\value{
A string representing the new file name with the version appended.
}
\description{
This internal function appends a form version string to a given file name,
using a specific delimiter pattern.
}
\examples{
\dontrun{
add_form_to_string("report.xlsx", "cp.v2.0.0")
}
}
\keyword{internal}
