% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_form_to_file_name.R
\name{extract_form_from_string}
\alias{extract_form_from_string}
\title{Extract form version from file name}
\usage{
extract_form_from_string(file_name_with_form)
}
\arguments{
\item{file_name_with_form}{The file name from which to extract the version.}
}
\value{
A string representing the extracted version, or NA if not found.
}
\description{
This internal function extracts the form version string from a file name,
assuming it follows a specific delimiter pattern.
}
\examples{
\dontrun{
extract_form_from_string("{(-cp.v2.0.0-)}report.xlsx")
}
}
\keyword{internal}
