% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{file_tibble}
\alias{file_tibble}
\title{Retrieve File Information for Multiple Paths}
\usage{
file_tibble(
  folder_paths,
  file_type = ".",
  recursive = FALSE,
  filter_out_tilda = TRUE
)
}
\arguments{
\item{folder_paths}{Character vector. Paths of the folders from which files' information is to be extracted.}

\item{file_type}{Character. Type or extension of the files to be considered. Default is "." (all files).}

\item{recursive}{Logical. Should files from subdirectories also be included? Default is FALSE.}

\item{filter_out_tilda}{Logical. Should files with names containing a tilde (~) be excluded? Default is TRUE.}
}
\value{
A tibble with one row per unique file-path combination, with columns: folder_paths (list of all folder paths provided),
file_type (type of the file considered), recursive (were subdirectories included?), filter_out_tilda (were files with tildes excluded?),
file (name of the file), path (complete path of the file), and metadata (a list column containing details like size, isdir, and mtime).
}
\description{
This function aggregates file information from multiple folder paths into a single tibble.
}
\examples{
\dontrun{
file_tibble(folder_paths = c("./dir1", "./dir2"))
}
}
