% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{file_tibble_single_path}
\alias{file_tibble_single_path}
\title{Retrieve File Information for a Single Path}
\usage{
file_tibble_single_path(
  folder_path,
  file_type = ".",
  recursive = FALSE,
  filter_out_tilda = TRUE
)
}
\arguments{
\item{folder_path}{Character. The path of the folder from which files' information is to be extracted.}

\item{file_type}{Character. Type or extension of the files to be considered. Default is "." (all files).}

\item{recursive}{Logical. Should files from subdirectories also be included? Default is FALSE.}

\item{filter_out_tilda}{Logical. Should files with names containing a tilde (~) be excluded? Default is TRUE.}
}
\value{
A tibble containing columns: file (name of the file), path (complete path of the file),
size (file size), isdir (is it a directory?), and mtime (last modification time).
}
\description{
This function returns a tibble with details about files in a given folder path.
}
\examples{
\dontrun{
file_tibble_single_path(folder_path = "./my_directory")
}
#' @keywords internal
}
