% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{get_raw_xlsx}
\alias{get_raw_xlsx}
\title{Add Data Column to file_tibble}
\usage{
get_raw_xlsx(df, sheets = NULL, sheets_regex = ".*")
}
\arguments{
\item{df}{A tibble generated by the \code{file_tibble} function.}

\item{sheets}{A character vector specifying which sheets to extract from the Excel files.
If NULL (default), all sheets are extracted.}

\item{sheets_regex}{A regular expression pattern to match against sheet names in the
Excel files. Defaults to ".*" to match all sheets.}
}
\value{
A tibble with the original data and an additional list column \code{raw_df} that
contains the data from the Excel files.
}
\description{
The \code{get_raw_xlsx} function takes as input a tibble generated by the \code{file_tibble}
function and returns it with an added column \code{raw_df}. This column contains the data
from all sheets of the corresponding Excel files.
}
\examples{
\dontrun{
# Assume `file_tibble` has been previously run to generate `original_tibble`
raw_data_tibble <- get_raw_xlsx(original_tibble)
}

}
