% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{reform_tib}
\alias{reform_tib}
\title{Reform a Tibble Using Form Metadata}
\usage{
reform_tib(df)
}
\arguments{
\item{df}{A tibble containing the data to be reformed.}
}
\value{
A tibble with a new column named \code{output} containing the reformed data.
}
\description{
This function uses metadata from a specified \code{column} to reform a tibble.
It utilizes \code{get_form_template} to extract specific data based on the form name and version,
and applies transformations to create a new tibble.
}
\examples{
\dontrun{
reform_tib(my_data, "raw_form_column")
}

}
\seealso{
\code{\link{get_form_template}} for extracting form metadata.
}
