% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_form_to_file_name.R
\name{subset_file_name}
\alias{subset_file_name}
\title{Extract the file name from a full file path and name}
\usage{
subset_file_name(file_path_and_name)
}
\arguments{
\item{file_path_and_name}{The full file path and name.}
}
\value{
A string representing just the file name.
}
\description{
This function extracts the file name from a full file path and name string.
}
\examples{
\dontrun{
subset_file_name("/home/user/documents/report.xlsx")
}
}
