% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_form_to_file_name.R
\name{subset_file_path}
\alias{subset_file_path}
\title{Extract the file path from a full file path and name}
\usage{
subset_file_path(file_path_and_name)
}
\arguments{
\item{file_path_and_name}{The full file path and name.}
}
\value{
A string representing the file path without the file name.
}
\description{
This function extracts the path part from a full file path and name string.
}
\examples{
\dontrun{
subset_file_path("/home/user/documents/report.xlsx")
}
}
