% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{update_file_tibble}
\alias{update_file_tibble}
\title{Update Metadata Tibble for Files}
\usage{
update_file_tibble(df)
}
\arguments{
\item{df}{A tibble generated by the \code{file_tibble} function containing file metadata and
file paths.}
}
\value{
A tibble updated with the current metadata for the files. This includes any new files,
removed files, or files that have seen metadata changes since the original tibble was
generated. The tibble will also contain an 'update' column that is marked TRUE for
files that have been added or changed.
}
\description{
The \code{update_file_tibble} function takes as input a tibble that was originally generated
by the \code{file_tibble} function. This function checks for any changes in the file metadata
(e.g., additions, removals, or updates of files in the specified directories) and updates
the tibble accordingly.
}
\details{
The function first unpacks the metadata of the given \code{df}. It then retrieves the necessary
parameters like \code{folder_paths}, \code{file_type}, \code{recursive}, and \code{filter_out_tilda}. A new tibble
is generated using the \code{file_tibble} function with these parameters.

The function identifies any new rows (i.e., files) that have been added since the original
tibble was generated. It also identifies rows that are present in both the old and new tibbles
but have different values for the \code{size}, \code{isdir}, or \code{mtime} metadata fields.

Based on this comparison, the function updates the 'update_needed_raw' and 'update_needed_r'
columns in the tibble. Rows corresponding to newly added or changed files
are marked TRUE in both 'update_' columns.
}
\examples{
\dontrun{
# Assume `file_tibble` has been previously run to generate `original_tibble`
updated_tibble <- update_file_tibble(original_tibble)
}

}
\references{
The function uses several functions from the 'tidyverse' and makes references to
other custom packages. Ensure the necessary dependencies are installed and loaded.
}
