% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{update_raw_xlsx}
\alias{update_raw_xlsx}
\title{Update Data from Excel Files Based on Metadata Changes}
\usage{
update_raw_xlsx(df)
}
\arguments{
\item{df}{A tibble generated by the \code{get_raw_xlsx} function.}
}
\value{
A tibble with the original data and updated \code{raw_df} column for any files with
changed metadata.
}
\description{
The \code{update_raw_xlsx} function takes a tibble generated by \code{get_raw_xlsx} as input.
It checks if any files in the tibble have updated metadata, and if so, updates the
\code{raw_df} column with the latest data from those files.
}
\details{
If the input tibble does not have a \code{raw_df} column, the function will raise an error.
}
\examples{
\dontrun{
# Assume `get_raw_xlsx` has been previously run to generate `raw_data_tibble`
updated_data_tibble <- udate_raw_xlsx(raw_data_tibble)
}

}
